Backend/requirements.txt
astroid==2.2.5
autopep8==1.4.4
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
Pillow==6.1.0
pkg-resources==0.0.0
psycopg2==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
